+++ src/test/extension.test.ts
@@ -1,9 +1,10 @@
 import * as assert from 'assert';
 
-// You can import and use all API from the 'vscode' module
-// as well as import your extension to test it
+
 import * as vscode from 'vscode';
-// import * as myExtension from '../../extension';
+import { Workspace } from '../workspace/Workspace';
+import { Todo } from '../Todo/Todo';
+import { Editor } from '../Editor/Editor';
 
 suite('Extension Test Suite', () => {
 	vscode.window.showInformationMessage('Start all tests.');